chore: refactor player components into dedicated directory
