+++ src/components/player/VideoPage.tsx
@@ -1,6 +1,6 @@
 import LikeButtons from "./LikeButtons.tsx";
 import VideoPlayer from "./VideoPlayer.tsx";
-import { Like, Movie } from "../types.ts";
+import { Like, Movie } from "../../types.ts";
 
 export default function VideoPage(
   { movie, like }: { movie: Movie; like?: Like },
